rename to ComponentPropsTable, add useComponentProps hook
