+++ docs/src/components/ComponentDoc/ComponentPropsTable/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ComponentPropsTable'
