--- docs/src/components/ComponentDoc/ComponentTable/ComponentTableHeader.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react'
-
-// TODO: use Flex or a Table component, when it will be available
-const ComponentTableHeader = () => (
-  <thead>
-    <tr>
-      <th>Name</th>
-      <th>Default</th>
-      <th>Type</th>
-      <th>Description</th>
-    </tr>
-  </thead>
-)
-
-const arePropsEqual = () => true
-
-export default React.memo(ComponentTableHeader, arePropsEqual)
