--- docs/src/components/ComponentDoc/ComponentTable/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ComponentTable'
