+++ docs/src/components/ComponentDoc/useComponentProps.ts
@@ -1,0 +1,15 @@
+import componentInfoContext from 'docs/src/utils/componentInfoContext'
+
+const useComponentProps = (componentName: string) => {
+  const info = componentInfoContext.byDisplayName[componentName]
+
+  if (!info) {
+    throw new Error(
+      `We don't have definitions for "${componentName}", please ensure that are generated by "gulp-docgen" plugin`,
+    )
+  }
+
+  return info.props
+}
+
+export default useComponentProps
